Add files via upload
Purpose of the application

This application is intended as a teaching aid, for the teaching of subjects dealing with signals and signals. The application is available in English and Slovak language. This application is created for the course Introduction to Signals and Signals , which is taught at the Multimedia and Information and Communication Technologies Department at the Faculty of Electrical Engineering and Information Technology at the University of Žilina. This application was created as my bachelor thesis.

Thesis in application: 
	Complex numbers
	Basic signal distribution
	Fourier series
	Fourier transform
	Discrete Fourier transform and sampling
	Convolution and correlation
	Stochastic signals
	Pseudo-random sequence
	Line codes
	Shannon-Hartley theorem (channel capacity)
	Analog frequency filters

Content in the application

This application contains for each topic a theoretical text part, tests, interactive demonstrations of signals,videos, practical tasks and examples with visualization using MATLAB.

Navigation in the application

Navigation is provided via buttons in the main menu or via tabs.
 
Theoretical text part

The theory in the Slovak version of the application is created as an HTML document. In the case of the English version the PDF format is used.
 
Tests

The tests in the app offer a choice of four options where only one option is correct. After clicking the submit button, the result of the test will be displayed as a percentage. The restart button can be used to restart the test and write it again.
 
Interactive examples

The app also includes interactive signal examples where the user can change the value, for example, using a slider and the waveform changes based on the user's input.
 
Examples

The examples in the application are designed in such a way that the user first calculates the example and draws the signal waveform or its spread based on the calculated result. In order to allow the user to write code, a simple code editor is provided in the application that allows automatic writing of variables and built-in functions. The shortcut to use this feature is CTRL+SPACE. The code editor was created using the open source project CodeMirror available at: https://codemirror.net/5/.The solution is also included in the attachments and can be viewed via the show answer button.
 
Other features

In addition to the above features, the application includes a PDF document with instructions on how to use the application, a table with basic MATLAB commands, solved examples with step-by-step instructions, and other additional examples.